moved connection code to seperate file
